requirements.txt
flask==3.0.0
requests==2.31.0
matplotlib==3.8.2
python-dotenv==1.0.0
numpy==1.26.2
openai>=1.0.0
